+++ action_server/src/sema4ai/action_server/vendored_deps/package_deps/_conda_deps.py
@@ -1,5 +1,5 @@
+from collections.abc import Iterator
 from dataclasses import dataclass
-from typing import Dict, Iterator, Optional
 
 from ..ls_protocols import _RangeTypedDict
 from .conda_impl import conda_match_spec, conda_version
@@ -9,11 +9,13 @@ from .conda_impl import conda_match_spec, conda_version
 class CondaDepInfo:
     name: str  # The name of the dep (i.e.: python)
     value: str  # The full value of the dep (i.e.: python=3.7)
-    version_spec: str  # The version SPEC of the dep (as seen by conda: '3.7.*' or '>3.2')
+    version_spec: (
+        str  # The version SPEC of the dep (as seen by conda: '3.7.*' or '>3.2')
+    )
     dep_range: _RangeTypedDict
-    error_msg: Optional[str] = None
+    error_msg: str | None = None
 
-    def get_dep_vspec(self) -> Optional[conda_version.VersionSpec]:
+    def get_dep_vspec(self) -> conda_version.VersionSpec | None:
         try:
             vspec = conda_version.VersionSpec(self.version_spec)
         except Exception:
@@ -23,7 +25,7 @@ class CondaDepInfo:
 
 class CondaDeps:
     def __init__(self) -> None:
-        self._deps: Dict[str, CondaDepInfo] = {}
+        self._deps: dict[str, CondaDepInfo] = {}
 
     def add_dep(self, value: str, dep_range: _RangeTypedDict):
         """
@@ -42,7 +44,7 @@ class CondaDeps:
         else:
             self._deps[name] = CondaDepInfo(name, value, version_spec, dep_range)
 
-    def get_dep_vspec(self, spec_name: str) -> Optional[conda_version.VersionSpec]:
+    def get_dep_vspec(self, spec_name: str) -> conda_version.VersionSpec | None:
         conda_dep_info = self._deps.get(spec_name)
         if conda_dep_info is None:
             return None